backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pydantic==2.5.0
python-dateutil==2.8.2
pydantic-settings==2.1.0
openpyxl==3.1.2
reportlab==4.0.7
pandas==2.1.3
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

